event ticket email additional content bug fixed: The bug was sending multiple additional content with each tickets as it was using a hook which gets trigger in a loop

+++ event-ticket-email-additional-content.php
@@ -61,13 +61,17 @@ if ( ! class_exists( '' ) ) {
 	class EventTicketEmailContent {
 
 		protected $event_id;
+		protected $alreadyAdded;
 
 
 		function __construct() {
+
+			$this->alreadyAdded = false;
+
 			add_action( 'add_meta_boxes', array($this, 'etpeac_metabox_create') );
 			add_action( 'save_post',  array($this, 'etpeac_meta_save'), 10, 3);
-			add_action( 'tribe_tickets_ticket_email_ticket_top', array($this,'etpeac_store_event_id') );
-			//add_action( 'tribe_tickets_ticket_email_bottom', array($this,'etpeac_add_additional_content') );
+			add_action( 'tribe_tickets_ticket_email_ticket_top', array($this,'etpeac_add_additional_content') );
+
 		}
 
 
@@ -92,8 +96,6 @@ if ( ! class_exists( '' ) ) {
 		 * @return  null
 		 */
 		public function etpeac_metabox_display($post){
-
-
 
 			// Ensure the global $post variable is in scope
 			global $post;
@@ -152,69 +154,42 @@ if ( ! class_exists( '' ) ) {
 
 		}
 
-		public function etpeac_store_event_id($ticket){
-			$this->event_id = $ticket['event_id'];
+		public function etpeac_add_additional_content($ticket){
 
-			$args = array(
-				'post_type'     => 'eemailcontent',
-				'meta_query' => array(
-					array(
-						'key' => 'event_id',
-						'value' =>  $this->event_id,
-						'compare' => '=',
+			if(!$this->alreadyAdded){
+				$this->event_id = $ticket['event_id'];
+				$args = array(
+					'post_type'     => 'eemailcontent',
+					'meta_query' => array(
+						array(
+							'key' => 'event_id',
+							'value' =>  $this->event_id,
+							'compare' => '=',
+						)
 					)
-				)
-			);
+				);
 
 
-			$query = new WP_Query($args);
-			echo '<table class="content" align="center" width="620" cellspacing="0" cellpadding="0" border="0" bgcolor="#ffffff" style="margin:0 auto; padding:0; margin: top -100px">';
-			echo '<tr>';
-			echo '<td style="text-align: left; color: #555;">';
-			if( $query->have_posts() ){
-				while ($query->have_posts()){
-					$query->the_post();
-					wpautop(the_content());
-					break;
+				$query = new WP_Query($args);
+				echo '<table class="content" align="center" width="620" cellspacing="0" cellpadding="0" border="0" bgcolor="#ffffff" style="margin:0 auto; padding:0; margin: top -100px">';
+				echo '<tr>';
+				echo '<td style="text-align: left; color: #555;">';
+				if( $query->have_posts() ){
+					while ($query->have_posts()){
+						$query->the_post();
+						wpautop(the_content());
+						break;
+					}
 				}
+				echo '</td>';
+				echo '</tr>';
+				echo '<table>';
+
+				$this->alreadyAdded = true;
+
 			}
-			echo '</td>';
-			echo '</tr>';
-			echo '<table>';
-		}
-
-
-		public function etpeac_add_additional_content(){
-
-			$args = array(
-				'post_type'     => 'eemailcontent',
-				'meta_query' => array(
-					array(
-						'key' => 'event_id',
-						'value' =>  $this->event_id,
-						'compare' => '=',
-					)
-				)
-			);
-
-
-			$query = new WP_Query($args);
-			echo '<table class="content" align="center" width="620" cellspacing="0" cellpadding="0" border="0" bgcolor="#ffffff" style="margin:20px auto; padding:0; margin: top -100px">';
-			echo '<tr>';
-			echo '<td style="text-align: left; color: #555;">';
-			if( $query->have_posts() ){
-				while ($query->have_posts()){
-					$query->the_post();
-					wpautop(the_content());
-					break;
-				}
-			}
-			echo '</td>';
-			echo '</tr>';
-			echo '<table>';
 
 		}
-
 
 
 	}
